Create an orthogonal array using the Bush algorithm.

The bush program produces OA( q^3, k, q, 3 ), k <= q+1 for prime powers q.

createBush(q, ncol, bRandom = TRUE)

Arguments

q

the number of symbols in the array

ncol

number of parameters or columns

bRandom

should the array be randomized

Value

an orthogonal array

Details

From Owen: An orthogonal array A is a matrix of n rows, k columns with every element being one of q symbols 0,...,q-1. The array has strength t if, in every n by t submatrix, the q^t possible distinct rows, all appear the same number of times. This number is the index of the array, commonly denoted lambda. Clearly, lambda*q^t=n. The notation for such an array is OA( n, k, q, t ).

References

Owen, Art. Orthogonal Arrays for: Computer Experiments, Visualizations, and Integration in high dimenstions. http://lib.stat.cmu.edu/designs/oa.c. 1994 K.A. Bush (1952) Annals of Mathematical Statistics, Vol 23 pp 426-434

See also

Examples

A <- createBush(3, 3, FALSE)
B <- createBush(4, 5, TRUE)