Optimum Latin Hypercube Sample

Draws a Latin Hypercube Sample from a set of uniform distributions for use in creating a Latin Hypercube Design. This function uses the Columnwise Pairwise (CP) algorithm to generate an optimal design with respect to the S optimality criterion.

optimumLHS(n = 10, k = 2, maxSweeps = 2, eps = 0.1, verbose = FALSE)

Arguments

n
The number of partitions (simulations or design points or rows)

k
The number of replications (variables or columns)

maxSweeps
The maximum number of times the CP algorithm is applied to all the columns.

eps
The optimal stopping criterion. Algorithm stops when the change in optimality measure is less than eps*100% of the previous value.

verbose
Print informational messages

Value

An n by k Latin Hypercube Sample matrix with values uniformly distributed on [0,1]

Details

Latin hypercube sampling (LHS) was developed to generate a distribution of collections of parameter values from a multidimensional distribution. A square grid containing possible sample points is a Latin square iff there is only one sample in each row and each column. A Latin hypercube is the generalisation of this concept to an arbitrary number of dimensions. When sampling a function of k variables, the range of each variable is divided into n equally probable intervals. n sample points are then drawn such that a Latin Hypercube is created. Latin Hypercube sampling generates more efficient estimates of desired parameters than simple Monte Carlo sampling.

This program generates a Latin Hypercube Sample by creating random permutations of the first n integers in each of k columns and then transforming those integers into n sections of a standard uniform distribution. Random values are then sampled from within each of the n sections. Once the sample is generated, the uniform sample from a column can be transformed to any distribution by using the quantile functions, e.g. qnorm(). Different columns can have different distributions.

S-optimality seeks to maximize the mean distance from each design point to all the other points in the design, so the points are as spread out as possible.

This function uses the CP algorithm to generate an optimal design with respect to the S optimality criterion.

References

Stocki, R. (2005) A method to improve design reliability using optimal Latin hypercube sampling Computer Assisted Mechanics and Engineering Sciences 12, 87--105.

See also

Examples

A <- optimumLHS(4, 3, 5, .05)